# docs/backend/openapi.yaml
openapi: 3.1.0
info:
  title: Micocards API
  version: 0.1.0
  summary: Backend API for the Micocards flashcard service.
  description: |
    REST + JSON. Auth via Authorization: Bearer <jwt>. Non-idempotent writes accept
    an Idempotency-Key header. Status codes follow ADR 0006.
  contact:
    name: Micocards backend
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://v811467.hosted-by-vdsina.com
    description: Production
...
paths:
...
  /api/decks:
    get:
      tags: [decks]
      operationId: listUserDecks
      summary: List the authenticated user's decks.
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - in: query
          name: cursor
          schema: { type: string }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeckList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/Internal" }
    post:
      tags: [decks]
      operationId: createDeck
      summary: Create a new deck (optionally with cards).
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateDeckRequest" }
      responses:
        "201":
          description: created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Deck" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/Unprocessable" }
        "500": { $ref: "#/components/responses/Internal" }
  /api/decks/generate:
    post:
      tags: [decks]
      operationId: generateDeckWithAI
      summary: Generate a deck draft with AI (stub; returns 501 unless AI_API_KEY is set).
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GenerateDeckRequest" }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AIDeckDraftResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "501": { $ref: "#/components/responses/NotImplemented" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "500": { $ref: "#/components/responses/Internal" }
  /api/decks/{deckID}:
    parameters:
      - $ref: "#/components/parameters/DeckID"
    get:
      tags: [decks]
      operationId: getDeck
      summary: Get a deck by id.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Deck" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/Internal" }
    patch:
      tags: [decks]
      operationId: renameDeck
      summary: Rename a deck.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RenameDeckRequest" }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Deck" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Unprocessable" }
        "500": { $ref: "#/components/responses/Internal" }
    delete:
      tags: [decks]
      operationId: deleteDeck
      summary: Soft-delete a deck.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/Internal" }
  /api/decks/{deckID}/cards:
    parameters:
      - $ref: "#/components/parameters/DeckID"
    post:
      tags: [cards]
      operationId: addCard
      summary: Append a card to the deck.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AddCardRequest" }
      responses:
        "201":
          description: created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Card" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Unprocessable" }
        "500": { $ref: "#/components/responses/Internal" }
  /api/decks/{deckID}/cards/{cardID}:
    parameters:
      - $ref: "#/components/parameters/DeckID"
      - $ref: "#/components/parameters/CardID"
    patch:
      tags: [cards]
      operationId: editCard
      summary: Edit a card.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EditCardRequest" }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Card" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/Unprocessable" }
        "500": { $ref: "#/components/responses/Internal" }
    delete:
      tags: [cards]
      operationId: removeCard
      summary: Remove a card.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/Internal" }
  /api/decks/{deckID}/cards/order:
    parameters:
      - $ref: "#/components/parameters/DeckID"
    put:
      tags: [cards]
      operationId: reorderCards
      summary: Replace the card ordering of a deck.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReorderCardsRequest" }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CardList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/Unprocessable" }
        "500": { $ref: "#/components/responses/Internal" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: false
      schema: { type: string, minLength: 1, maxLength: 128 }
      description: Required for non-idempotent writes per ADR 0005.
    DeckID:
      in: path
      name: deckID
      required: true
      schema: { type: string, format: uuid }
    CardID:
      in: path
      name: cardID
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: bad request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: unauthorized
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: forbidden
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: conflict
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unprocessable:
      description: unprocessable entity
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotImplemented:
      description: not implemented
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadGateway:
      description: upstream failure
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Internal:
      description: internal server error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Ok:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean, enum: [true] }
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
        message: { type: string }
        details: { type: object, additionalProperties: true }
...
    Card:
      type: object
      required: [id, term, definition, ordinal]
      properties:
        id: { type: string, format: uuid }
        term: { type: string }
        definition: { type: string }
        ordinal: { type: integer, minimum: 1 }
    CardList:
      type: object
      required: [cards]
      properties:
        cards:
          type: array
          items: { $ref: "#/components/schemas/Card" }
    CardDraft:
      type: object
      required: [term, definition]
      properties:
        term: { type: string, minLength: 1, maxLength: 512 }
        definition: { type: string, minLength: 1, maxLength: 2048 }
    Deck:
      type: object
      required: [id, ownerId, title, cards, createdAt]
      properties:
        id: { type: string, format: uuid }
        ownerId: { type: string, format: uuid }
        title: { type: string }
        cards:
          type: array
          items: { $ref: "#/components/schemas/Card" }
        createdAt: { type: string, format: date-time }
    DeckSummary:
      type: object
      required: [id, title, cardCount, createdAt]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        cardCount: { type: integer, minimum: 0 }
        createdAt: { type: string, format: date-time }
    DeckList:
      type: object
      required: [decks]
      properties:
        decks:
          type: array
          items: { $ref: "#/components/schemas/DeckSummary" }
        nextCursor: { type: string }
    CreateDeckRequest:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1, maxLength: 120 }
        cards:
          type: array
          maxItems: 500
          items: { $ref: "#/components/schemas/CardDraft" }
    RenameDeckRequest:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1, maxLength: 120 }
    AddCardRequest:
      $ref: "#/components/schemas/CardDraft"
    EditCardRequest:
      type: object
      properties:
        term: { type: string, minLength: 1, maxLength: 512 }
        definition: { type: string, minLength: 1, maxLength: 2048 }
    ReorderCardsRequest:
      type: object
      required: [orderedIds]
      properties:
        orderedIds:
          type: array
          items: { type: string, format: uuid }
    GenerateDeckRequest:
      type: object
      required: [prompt]
      properties:
        prompt: { type: string, minLength: 1, maxLength: 1024 }
    AIDeckDraftResponse:
      type: object
      required: [draft]
      properties:
        draft:
          type: object
          required: [title, cards]
          properties:
            title: { type: string }
            cards:
              type: array
              items: { $ref: "#/components/schemas/CardDraft" }
...